api/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncio==3.4.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
fastapi==0.115.5
fastembed==0.4.2
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-ollama==0.2.0
langchain-postgres==0.0.12
langchain-text-splitters==0.3.2
langsmith==0.1.144
loguru==0.7.2
marshmallow==3.23.1
mmh3==4.1.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
ollama==0.4.0
onnx==1.17.0
onnxruntime==1.19.2
orjson==3.10.11
packaging==24.2
pgvector==0.2.5
pillow==10.4.0
propcache==0.2.0
protobuf==5.28.3
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
py_rust_stemmers==0.1.3
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
pyreadline3==3.5.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.3
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.20.3
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
win32-setctime==1.1.0
yarl==1.18.0

#Auth
PyJWT==2.10.1